python/app/requirements.txt
blis==0.2.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cymem==2.0.2
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.2
en-core-web-sm==2.1.0
firebase-admin==2.18.0
Flask==1.1.1
future==0.17.1
gcloud==0.17.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
msgpack==0.6.1
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.2
oauth2client==3.0.0
plac==0.9.6
preshed==2.0.1
protobuf==3.10.0rc1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2019.2
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
serializers==0.2.4
six==1.12.0
spacy==2.1.8
sqlparse==0.3.0
srsly==0.1.0
stop-words==2018.7.23
thinc==7.0.8
tqdm==4.35.0
uritemplate==3.0.0
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.5
